feat: modularize seedlot table further
