+++ frontend/src/views/Seedlot/MySeedlots/constants.ts
@@ -1,10 +1,7 @@
 export const tableText = {
   pageTitle: 'My Seedlots',
   pageSubtitle: 'Check and manage your own seedlots',
-  buttonText: 'Register a new seedlot',
-  emptyPictogram: 'Magnify',
-  emptyTitle: 'There is no seedlot to show yet!',
-  emptyDescription: 'Your seedlots will appear here once you generate one'
+  buttonText: 'Register a new seedlot'
 };
 
 export const headerData = [
