+++ frontend/src/views/Seedlot/MySeedlots/styles.scss
@@ -20,31 +20,6 @@
     float: right;
   }
 
-  .#{vars.$bcgov-prefix}--data-table-container {
-    width: 100%;
-  }
-
-  .my-seedlot-data-table-row {
-    margin: 0 -2rem;
-  }
-
-  .#{vars.$bcgov-prefix}--table-sort {
-    background-color: var(--#{vars.$bcgov-prefix}-layer-01);
-  }
-
-  .#{vars.$bcgov-prefix}--search-input {
-    border-bottom: 1px solid var(--#{vars.$bcgov-prefix}-border-subtle-01)
-  }
-
-  .#{vars.$bcgov-prefix}--data-table tr {
-    cursor: pointer;
-  }
-
-  .#{vars.$bcgov-prefix}--pagination__button,
-  .#{vars.$bcgov-prefix}--btn--ghost.#{vars.$bcgov-prefix}--pagination__button {
-    height: 3rem;
-  }
-
   .empty-seedlots{
     width: 100%;
     margin: 0 auto;
