+++ frontend/src/views/Seedlot/SeedlotDashboard/RecentSeedlots/constants.ts
@@ -1,12 +1,6 @@
 const recentSeedlotsText = {
   tableTitle: 'My seedlots',
-  tableSubtitle: 'Check a summary of your recent seedlots',
-  emptyPictogram: 'Magnify',
-  emptyTitle: 'There is no seedlot to show yet!',
-  emptyDescription: 'Your recent seedlots will appear here once you generate one',
-  errorTitle: 'Something went wrong...',
-  errorDescription: 'Could not display seedlot data',
-  errorIcon: 'InProgressError'
+  tableSubtitle: 'Check a summary of your recent seedlots'
 };
 
 export default recentSeedlotsText;
